requirements.txt
openai
langchain
python-dotenv
pyyaml
tqdm
rich
nltk
pdfkit